requirements.txt
# Создаем requirements.txt
# Core Dependencies
python>=3.10

# Deep Learning & NLP
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

# Vector Search
faiss-cpu==1.7.4

# Traditional IR
rank-bm25==0.2.2

# Data Processing
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0

# Text Processing (Russian)
razdel==0.5.0
pymorphy3==1.3.0

# Utils
tqdm==4.66.1
pyyaml==6.0
loguru==0.7.0

# Testing & Quality
pytest==7.4.0
black==23.7.0
ruff==0.0.285

# Experiment Tracking
mlflow==2.7.0
